2013-05-14  Sancho Lerena <slerena@artica.es>
        * include/config_process.php,
        integria.centos.spec: Version updates.

        * install.php: removed unused dependency (mcrypt).

        * integria_db.sql: Fixed a bad definition in tinventory_track table.

+++ integria_db.sql
@@ -594,7 +594,7 @@ CREATE TABLE `tmanufacturer` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 CREATE TABLE `tinventory` (
-  `id` mediumint(8) unsigned NOT NULL auto_increment,
+  `id` bigint(20) unsigned NOT NULL auto_increment,
   `id_object_type` mediumint(8) unsigned,
   `owner` varchar(60),
   `name` TEXT default '',
